# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Five 网页验证版接口
  version: 1.0.0
  summary: 公开网页、单人后台与后端共同使用的 P0 接口契约
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    本文件是 Five P0 接口字段、状态、错误码和版本保护规则的唯一事实源。
    `contentVersion`、各种 ID 和修订号都是不透明值，调用方不得解析其字符串。
    P0 只有网页端，因此接口不包含 `platform`、`sourcePlatform` 或 `targetPlatform`。
    `channelId` 只用于区分链接分发来源，不得改变日期、档位、正文或图片。
    所有 `/admin/api/v1` 响应（包括错误）必须使用 `Cache-Control: no-store`，防止退出后从缓存读取后台内容或核对凭证。
servers:
  - url: /
    description: 与网页同源的后端
...
security: []
paths:
...
  /admin/api/v1/auth/sessions:
    post:
      operationId: createAdminSession
      tags: [Admin authentication]
      summary: 使用账号与密码建立后台会话
      description: |
        服务端执行来源级和账号级持久化限流，并对不存在账号执行等价密码哈希工作量。
        校验成功后只通过同源 `HttpOnly`、`Secure`、`SameSite=Strict`、`Path=/admin`
        且无 `Domain` 的 Cookie 返回随机会话令牌，响应正文只返回与该会话绑定的 CSRF
        令牌和会话期限。此请求在会话建立前没有 CSRF 令牌，但仍必须通过可信同源
        `Origin` 校验。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminSessionRequest"
      responses:
        "201":
          description: 已建立 30 分钟闲置、12 小时绝对上限的后台会话
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            Set-Cookie:
              $ref: "#/components/headers/AdminSessionSetCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "400":
          $ref: "#/components/responses/AuthInvalidArgument"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "429":
          $ref: "#/components/responses/AuthRateLimited"
        "503":
          $ref: "#/components/responses/AdminServiceUnavailable"
...
  /admin/api/v1/operations/overview:
    get:
      operationId: getAdminOperationsOverview
      tags: [Admin operations]
      summary: 查看当前 servedFortuneDate 用户正在看到的内容和下一公开交付日准备状态
      security:
        - AdminSession: []
      responses:
        "200":
          description: 返回同一次服务端业务时间下的当前与下一公开交付日运营视图
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOperationsOverview"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /admin/api/v1/operations/calendar:
    get:
      operationId: getAdminOperationsCalendar
      tags: [Admin operations]
      summary: 查看由服务端日期关系生成的 42 格月历
      security:
        - AdminSession: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$"
          example: "2026-08"
      responses:
        "200":
          description: 返回月历窗口和统一日级摘要
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCalendarMonth"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /admin/api/v1/operations/issues:
    get:
      operationId: listAdminOperationsIssues
      tags: [Admin operations]
      summary: 查看维护者现在可以处理的问题
      description: 可选图片未请求、缺失或失败不会出现在此列表。
      security:
        - AdminSession: []
      responses:
        "200":
          description: 返回人类语言的可行动异常，不返回堆栈或内部错误对象
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionableIssueList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /admin/api/v1/operations/days/{fortuneDate}:
    get:
      operationId: getAdminOperationsDay
      tags: [Admin operations]
      summary: 查看某个命理日的真实用户端投影和订正能力
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/FortuneDate"
      responses:
        "200":
          description: 返回日级摘要、共享公开展示投影与并发保护信息
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDayDetail"
        "400":
          $ref: "#/components/responses/InvalidFortuneDate"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: five_admin_session
      description: |
        Issue #31 已确认的同源安全会话。生产 Cookie 固定为 `HttpOnly`、`Secure`、
        `SameSite=Strict`、`Path=/admin` 且无 `Domain`；闲置 30 分钟、绝对 12 小时失效。
        所有后台写操作还必须校验可信同源 `Origin` 与会话绑定的 `X-CSRF-Token`。
  parameters:
    FortuneDate:
      name: fortuneDate
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/FortuneDate"
...
  headers:
...
    XRequestId:
      description: 当前外部请求的追踪编号；缓存命中时也必须按请求生成。
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    NoStoreCacheControl:
      description: 认证、恢复、安全记录和紧急控制响应不得被浏览器或中间层保存。
      schema:
        type: string
        const: no-store
    AdminSessionSetCookie:
      description: |
        只包含随机会话令牌；生产环境固定使用 `HttpOnly; Secure; SameSite=Strict; Path=/admin`，
        不设置 `Domain`，闲置 30 分钟失效且绝对有效期不超过 12 小时。
      schema:
        type: string
...
  schemas:
    FortuneDate:
      type: string
      format: date
      description: Five 命理日；北京时间 23:00 进入下一命理日。
      example: "2026-07-15"
    ZonedDateTime:
      type: string
      format: date-time
      example: "2026-07-14T23:30:00+08:00"
...
    AdminUsername:
      type: string
      minLength: 3
      maxLength: 64
      pattern: '^[a-z0-9][a-z0-9._-]{2,63}$'
      description: 由离线 bootstrap 命令首次创建的唯一小写 ASCII 维护者账号名；限流与校验使用同一规范值。
    AdminUsernameInput:
      type: string
      minLength: 3
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{2,63}$'
      description: 登录输入；服务端在限流与密码校验前执行 NFKC、去除首尾空白并转为规范小写账号名。
    AdminPassword:
      type: string
      minLength: 8
      maxLength: 128
      writeOnly: true
      description: 以 Unicode code point 计 8 至 128 位；只从请求正文或交互式 stdin 接收，禁止进入 CLI 参数、环境变量或日志。
    CreateAdminSessionRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          $ref: "#/components/schemas/AdminUsernameInput"
        password:
          $ref: "#/components/schemas/AdminPassword"
    AdminSession:
      type: object
      additionalProperties: false
      required:
        - username
        - issuedAt
        - idleExpiresAt
        - absoluteExpiresAt
        - csrfToken
        - credentialRevision
      properties:
        username:
          $ref: "#/components/schemas/AdminUsername"
        issuedAt:
          $ref: "#/components/schemas/ZonedDateTime"
        idleExpiresAt:
          $ref: "#/components/schemas/ZonedDateTime"
        absoluteExpiresAt:
          $ref: "#/components/schemas/ZonedDateTime"
        csrfToken:
          type: string
          minLength: 32
          maxLength: 256
          description: 会话绑定的随机令牌，只保存在运行中页面内存。
        credentialRevision:
          type: integer
          minimum: 1
...
    RequestContext:
      type: object
      additionalProperties: false
      required:
        - responseGeneratedAt
        - civilDate
        - fortuneDate
        - shichen
        - timezone
        - dayBoundary
        - crossedDayBoundary
      properties:
        responseGeneratedAt:
          $ref: "#/components/schemas/ZonedDateTime"
        civilDate:
          type: string
          format: date
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        shichen:
          type: string
          enum: [子, 丑, 寅, 卯, 辰, 巳, 午, 未, 申, 酉, 戌, 亥]
        timezone:
          type: string
          const: Asia/Shanghai
        dayBoundary:
          type: string
          const: "23:00"
        crossedDayBoundary:
          type: boolean
    PublicContentContext:
      type: object
      additionalProperties: false
      required: [servedFortuneDate, switchBoundary, advancedFromCivilDate]
      description: |
        匿名用户和普通后台当前实际使用的公开内容日期。它只决定内容选择，
        不改变 23:00 命理日、时辰或日历算法。
      properties:
        servedFortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        switchBoundary:
          type: string
          const: "18:00"
        advancedFromCivilDate:
          type: boolean
...
    AdminDaySummary:
      type: object
      additionalProperties: false
      required:
        - fortuneDate
        - relation
        - effectiveFrom
        - effectiveTo
        - prepareBy
        - dayElement
        - dayElementLabel
        - primaryColors
        - operationalStatus
        - requiredImages
        - optionalImageStatus
        - previewAvailable
        - issueCodes
        - lifecycleRevision
        - scheduleSlotRevision
        - updatedAt
      properties:
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        relation:
          $ref: "#/components/schemas/AdminDayRelation"
        effectiveFrom:
          $ref: "#/components/schemas/ZonedDateTime"
        effectiveTo:
          $ref: "#/components/schemas/ZonedDateTime"
        prepareBy:
          $ref: "#/components/schemas/ZonedDateTime"
        dayElement:
          $ref: "#/components/schemas/ElementCode"
        dayElementLabel:
          type: string
          enum: [木, 火, 土, 金, 水]
        primaryColors:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/ColorRef"
        operationalStatus:
          $ref: "#/components/schemas/AdminOperationalStatus"
        requiredImages:
          $ref: "#/components/schemas/RequiredImageReadiness"
        optionalImageStatus:
          $ref: "#/components/schemas/OptionalImageStatus"
        previewAvailable:
          type: boolean
        issueCodes:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/AdminIssueCode"
        lifecycleRevision:
          type: integer
          minimum: 0
        scheduleSlotRevision:
          type: integer
          minimum: 0
        updatedAt:
          oneOf:
            - $ref: "#/components/schemas/ZonedDateTime"
            - type: "null"
    AdminOperationsOverview:
      type: object
      additionalProperties: false
      required:
        - requestContext
        - publicContentContext
        - nextOperationalBoundaryAt
        - health
        - issueCount
        - current
        - currentPreview
        - currentPreviewRequestContext
        - currentPreviewPublicContentContext
        - next
        - nextPreview
        - nextPreviewRequestContext
        - nextPreviewPublicContentContext
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        publicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
        nextOperationalBoundaryAt:
          $ref: "#/components/schemas/ZonedDateTime"
          description: 下一次运营刷新边界，只能是下一公开内容日的 prepareBy 或 18:00 公开生效时刻，不表示 23:00 命理换日或普通时辰边界。
        health:
          type: string
          enum: [healthy, attention, unavailable]
        issueCount:
          type: integer
          minimum: 0
        current:
          $ref: "#/components/schemas/AdminDaySummary"
        currentPreview:
          oneOf:
            - $ref: "#/components/schemas/DailyContent"
            - type: "null"
        currentPreviewRequestContext:
          $ref: "#/components/schemas/RequestContext"
        currentPreviewPublicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
          description: 服务端基于 currentPreviewRequestContext 解析的真实公开交付上下文；当前预览必须与本响应 publicContentContext 一致。
        next:
          $ref: "#/components/schemas/AdminDaySummary"
        nextPreview:
          oneOf:
            - $ref: "#/components/schemas/DailyContent"
            - type: "null"
        nextPreviewRequestContext:
          $ref: "#/components/schemas/RequestContext"
        nextPreviewPublicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
          description: 服务端基于 nextPreviewRequestContext 解析的预览公开交付上下文，客户端不得自行假定 advancedFromCivilDate。
    AdminCalendarMonth:
      type: object
      additionalProperties: false
      required: [requestContext, publicContentContext, nextOperationalBoundaryAt, month, items]
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        publicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
        nextOperationalBoundaryAt:
          $ref: "#/components/schemas/ZonedDateTime"
          description: 下一次运营刷新边界，只能是下一公开内容日的 prepareBy 或 18:00 公开生效时刻，不表示 23:00 命理换日或普通时辰边界。
        month:
          type: string
          pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$"
        items:
          type: array
          minItems: 42
          maxItems: 42
          items:
            $ref: "#/components/schemas/AdminDaySummary"
...
    AdminActionableIssueList:
      type: object
      additionalProperties: false
      required: [requestContext, publicContentContext, nextOperationalBoundaryAt, items]
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        publicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
        nextOperationalBoundaryAt:
          $ref: "#/components/schemas/ZonedDateTime"
          description: 下一次运营刷新边界，只能是下一公开内容日的 prepareBy 或 18:00 公开生效时刻，不表示 23:00 命理换日或普通时辰边界。
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminActionableIssue"
...
    AdminDayDetail:
      type: object
      additionalProperties: false
      required:
        - requestContext
        - publicContentContext
        - nextOperationalBoundaryAt
        - summary
        - previewSource
        - preview
        - previewRequestContext
        - previewPublicContentContext
        - editableSelectionKeys
        - readonlySelectionKeys
        - concurrency
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        publicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
        nextOperationalBoundaryAt:
          $ref: "#/components/schemas/ZonedDateTime"
          description: 下一次运营刷新边界，只能是下一公开内容日的 prepareBy 或 18:00 公开生效时刻，不表示 23:00 命理换日或普通时辰边界。
        summary:
          $ref: "#/components/schemas/AdminDaySummary"
        previewSource:
          type: string
          enum: [published, scheduled, approved, draft, none]
        preview:
          oneOf:
            - $ref: "#/components/schemas/DailyContent"
            - type: "null"
        previewRequestContext:
          $ref: "#/components/schemas/RequestContext"
        previewPublicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
          description: 服务端基于 previewRequestContext 解析的预览公开交付上下文，客户端不得自行计算 servedFortuneDate 或 advancedFromCivilDate。
        editableSelectionKeys:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        readonlySelectionKeys:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        concurrency:
          $ref: "#/components/schemas/AdminDayConcurrency"
...
    DailyContent:
      type: object
      additionalProperties: false
      required:
        - fortuneDate
        - effectiveFrom
        - effectiveTo
        - calendar
        - tiers
        - balanceSuggestion
        - outfitFormulas
        - looks
        - basis
        - share
        - versions
      properties:
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        effectiveFrom:
          $ref: "#/components/schemas/ZonedDateTime"
        effectiveTo:
          $ref: "#/components/schemas/ZonedDateTime"
        calendar:
          $ref: "#/components/schemas/CalendarInfo"
        tiers:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: "#/components/schemas/Tier"
        balanceSuggestion:
          $ref: "#/components/schemas/BalanceSuggestion"
        outfitFormulas:
          type: array
          minItems: 3
          items:
            $ref: "#/components/schemas/OutfitFormula"
        looks:
          type: array
          minItems: 2
          maxItems: 3
          description: 恰好两项 `requiredForPublish=true`，第三项如果存在必须为可选。
          items:
            $ref: "#/components/schemas/PublicLook"
        basis:
          $ref: "#/components/schemas/Basis"
        share:
          $ref: "#/components/schemas/ShareContent"
        versions:
          $ref: "#/components/schemas/ContentVersionSet"
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
  responses:
    InvalidArgument:
      description: 请求字段或格式无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AuthInvalidArgument:
      description: 认证、恢复或高风险确认请求字段无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthenticationFailed:
      description: 账号不存在或密码错误时使用相同状态和对外文案，不能帮助枚举账号。
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidFortuneDate:
      description: 命理日格式或允许范围无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthenticated:
      description: 后台会话不存在或已失效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AdminServiceUnavailable:
      description: 后台凭据库或安全控制服务暂时不可用；调用方应保留当前输入并稍后重试
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    CsrfValidationFailed:
      description: 后台写请求的可信同源 `Origin` 或会话绑定 CSRF 令牌校验失败
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AuthRateLimited:
      description: 登录或恢复请求触发持久化的请求来源级或账号级限流
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...